// src/Types/formTypes.ts
export type FormPrimaryValue = {
  name: string;
  surName: string;
  phone: string;
  email: string;
  startDate: null | string;
  endDate: null | string;
};

export type FormSecondaryValue = {
  name: string;
  phone: string;
};

export type FormLoginAndSignup = {
  email: string;
  password: string;
};